prod version maby

+++ backend/controllers/ListController.js
@@ -1,16 +1,25 @@
-const {List} = require('../models/models')
+const {User} = require("../models/models");
 
+const {List} = require('../models/models')
 
 class ListController {
 
     async create(req, res) {
-        const {text} = req.body
-        const obj = await List.findOne({where: {text}})
-        if (obj) {
-            return res.json({result_code: 1, message: "exist"})
+        try {
+            const {text, userId} = req.body
+            if (!text || !userId) {
+                return res.status(401).json({result_code: 1, message: "request error (paste text, userId)"})
+            }
+            const obj = await List.findOne({where: {text, userId}})
+            if (obj) {
+                return res.status(401).json({result_code: 1, message: "list exist"})
+            }
+            const list = await List.create({text, userId})
+
+            return res.json(list)
+        } catch (e) {
+            return res.status(401).json({result_code: 1, message: "ERROR"})
         }
-        const list = await List.create({text})
-        return res.json(list)
     }
 
     async update(req, res) {
@@ -22,12 +31,29 @@ class ListController {
 
     async deleteList(req, res) {
         const {id} = req.body
-        const list = await List.destroy({where: {id: id}})
+        const user = req.user
+        const list = await List.findOne({where: {id: id, userId: user.id}})
+        if (!list) {
+            return res.status(401).json({result_code: 1, message: "list not found"})
+        }
+        await List.destroy({where: {id: id}})
         return res.json(list)
     }
 
-    async getAll(req, res) {
+    async getAllDev(req, res) {
         const lists = await List.findAll()
+        return res.json(lists)
+    }
+
+    async getAll(req, res) {
+        const {userId} = req.query
+        if (!userId || userId === "null") {
+            return res.status(400).json({result_code: 1, message: "incorrect request (paste userId)"})
+        }
+        const lists = await List.findAll({where: {userId: req.query.userId}})
+        if (!lists || lists.length === 0) {
+            return res.json({result_code: 1, message: "list not exist"})
+        }
         return res.json(lists)
     }
